double

+++ main.cpp
@@ -2,7 +2,7 @@
 using namespace pxt;
 namespace tools {
 
-   int circleArea(int radius)
+   double circleArea(double radius)
    {
     return 3.14*radius*radius;
    }
